+++ Assets/Scripts/scene-shaping/ClickTarget.cs.meta
