+++ Assets/Structures/shaping.meta
