+++ Assets/Structures/shaping/泥6.fbx.meta
