+++ Assets/Structures/shaping/泥４.fbx.meta
